+++ src/client/test_socket.js
@@ -1,9 +1,14 @@
 'use strict';
 
-import io from 'socket.io-browserify';
-var socket = io.connect()
+import io from 'socket.io-client';
+var socket = io();
 
-socket.emit('get_articles', { 'module': 'hero' });
+socket.emit('get_popular');
+socket.on('got_popular', function(data) {
+  console.log(data);
+});
+
+/*socket.emit('get_articles', { 'module': 'hero' });
 socket.emit('get_articles', { 'source': 'freep' });
 
 socket.on('got_articles', function(data) {
@@ -14,4 +19,4 @@ socket.on('got_articles', function(data) {
 socket.on('new_articles', function(data) {
   console.log('New articles!');
   console.log(data);
-});
+});*/