+++ src/server/lib/route.js
@@ -1,0 +1,7 @@
+'use strict';
+
+export default function Catch(fn) {
+  return function(req, res, next) {
+    fn(req, res, next).catch(next)
+  };
+}