+++ src/server/routes/index.js
@@ -3,7 +3,10 @@
 import { Router } from 'express';
 
 import v1 from './v1';
-import { Article } from '../db';
+import { Article, Toppages } from '../db';
+import { Catch } from '../lib/index';
+import debug from 'debug';
+var logger = debug('app:route');
 
 var router = Router();
 
@@ -15,17 +18,28 @@ router.get('/test_socket/', function(req, res, next) {
   res.render('test_socket');
 });
 
-var socket = {
-  articles: function(app) {
-    return app.io.route('get_articles', async function(req) {
-      let articles;
-      try {
-        articles = await Article.find(req.data).exec();
-      } catch (err) {
-        logger.error(err);
-      }
-      req.io.emit('got_articles', { articles, filters: req.data });
-    });
-  }
-};
-module.exports = { index: router, v1, socket };
+router.get('/popular/', Catch(async function(req, res, next) {
+  let snapshot = await getPopular().exec();
+  req.io.emit('got_popular', { snapshot });
+  res.json({ success: true });
+}));
+
+function articles(socket) {
+  socket.on('get_articles', Catch(async function(req, res, next) {
+    let articles = await Article.find(req.data).exec();
+    socket.emit('got_articles', { articles, filters: req.data });
+  }));
+}
+
+function popular(socket) {
+  socket.on('get_popular', Catch(async function() {
+    let snapshot = await getPopular().exec();
+    socket.emit('got_popular', { snapshot });
+  }));
+}
+
+function getPopular() {
+  return Toppages.findOne().sort({ _id: -1 });
+}
+
+module.exports = { index: router, v1, popular, articles };